Clean up code
- Add extension methods
- Add tests
