+++ src/point.ts
@@ -1,0 +1,11 @@
+export { Point };
+
+class Point {
+    constructor(x: number, y: number) {
+        this.x = x;
+        this.y = y;
+    }
+
+    readonly x: number;
+    readonly y: number;
+}
